+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `WriteOptions::list_separator`
+- Implemented `WriteBuffer` and `ToStringWithOptions` for `NumberList` and `LengthList`. 
 
 ## [0.9.1] - 2017-12-15
 ### Fixed